+++ frontend/App.java
@@ -1,7 +1,5 @@
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
 
 public class App {
     public static void main(String[] args) {
@@ -40,50 +38,5 @@ public class App {
         frame.pack();
         frame.setLocationRelativeTo(null);
         frame.setVisible(true);
-    }
-}
-
-class Block extends JPanel {
-    private boolean colored = false;
-
-    Block() {
-        setPreferredSize(new Dimension(80, 80));
-        setBorder(BorderFactory.createLineBorder(Color.BLACK));
-        addMouseListener(new MouseAdapter() {
-            @Override
-            public void mouseClicked(MouseEvent e) {
-                colored = !colored;
-                setBackground(colored ? Color.BLUE : Color.GRAY);
-            }
-        });
-    }
-
-    @Override
-    public Insets getInsets() {
-        return new Insets(0, 0, 0, 0);
-    }
-}
-
-class Cursor extends JPanel {
-    private int x;
-    private int y;
-
-    Cursor() {
-        setPreferredSize(new Dimension(30, 30));
-        setOpaque(false); // Remove the background color
-        addMouseListener(new MouseAdapter() {
-            @Override
-            public void mouseMoved(MouseEvent e) {
-                x = e.getX();
-                y = e.getY();
-                setLocation(x - 15, y - 15);
-                repaint();
-            }
-        });
-    }
-
-    @Override
-    public Insets getInsets() {
-        return new Insets(0, 0, 0, 0);
     }
 }
