+++ frontend/Block.java
@@ -1,0 +1,29 @@
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
+
+class Block extends JPanel {
+    private boolean colored = false;
+
+    Block() {
+        setPreferredSize(new Dimension(80, 80));
+        setBorder(BorderFactory.createLineBorder(Color.BLACK));
+        addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                if (colored == true) {
+                    return;
+                }
+
+                colored = !colored;
+                setBackground(colored ? Color.BLUE : Color.GRAY);
+            }
+        });
+    }
+
+    @Override
+    public Insets getInsets() {
+        return new Insets(0, 0, 0, 0);
+    }
+}
